backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
strawberry-graphql==0.211.1
pydantic==2.4.2
aiosqlite==0.19.0
pytest==7.4.3
httpx==0.25.0
requests==2.31.0
python-dotenv==1.0.0
